Added Database+Hooks.swift
